Do not remove project file is build is in progress

+++ internal/runners/packages/packages.go
@@ -17,6 +17,7 @@ import (
 	"github.com/ActiveState/cli/internal/runbits"
 	"github.com/ActiveState/cli/pkg/platform/api/inventory/inventory_client/inventory_operations"
 	"github.com/ActiveState/cli/pkg/platform/model"
+	"github.com/ActiveState/cli/pkg/platform/runtime/artifact"
 	"github.com/ActiveState/cli/pkg/project"
 	"github.com/ActiveState/cli/pkg/projectfile"
 	"github.com/go-openapi/strfmt"
@@ -50,7 +51,7 @@ func executePackageOperation(prime primeable, packageName, packageVersion string
 			return locale.WrapError(err, "err_package_get_project", "Could not get project from path")
 		}
 		defer func() {
-			if rerr != nil {
+			if rerr != nil && !errors.Is(err, artifact.CamelRuntimeBuilding) {
 				if err := os.Remove(pj.Source().Path()); err != nil {
 					logging.Error("could not remove temporary project file: %s", errs.JoinMessage(err))
 				}
